initial redux support
