+++ package.json
@@ -7,9 +7,14 @@
 		"test": "jest"
 	},
 	"dependencies": {
+		"babel-plugin-transform-decorators-legacy": "^1.3.4",
 		"moment": "^2.18.1",
 		"react": "16.0.0-alpha.12",
-		"react-native": "0.47.1"
+		"react-native": "0.47.1",
+		"react-redux": "^5.0.6",
+		"redux": "^3.7.2",
+		"redux-persist": "^4.8.3",
+		"redux-thunk": "^2.2.0"
 	},
 	"devDependencies": {
 		"babel-jest": "20.0.3",
