+++ reducers/actionTypes.js
@@ -1,0 +1,2 @@
+export const SET_PAGE_DATA = 'SET PAGE DATA';
+export const UPDATE_BOOK = 'UPDATE_BOOK';
